first font reversed2

+++ ReverseFont/bitoperations.h
@@ -13,7 +13,7 @@
 
 uint8_t reverseBits(uint8_t num);
 
-uint8_t setBit(uint8_t &byte, uint8_t bit);
+uint8_t setBit(uint16_t &byte, uint8_t bit);
 
 uint8_t clearBit(uint8_t &byte, uint8_t bit);
